+++ DemoAppSwiftUI/LoginViewModel.swift
@@ -25,7 +25,6 @@ import SwiftUI
     private func connectUser(withCredentials credentials: UserCredentials) {
         loading = true
         let token = try! Token(rawValue: credentials.token)
-        LogConfig.level = .warning
 
         chatClient.connectUser(
             userInfo: .init(
@@ -50,7 +49,6 @@ import SwiftUI
 
     private func connectGuestUser(withCredentials credentials: UserCredentials) {
         loading = true
-        LogConfig.level = .warning
 
         chatClient.connectGuestUser(
             userInfo: .init(id: credentials.id, name: credentials.name)