+++ Sources/StreamChatSwiftUI/ChatChannel/ChatChannelView.swift
@@ -149,7 +149,7 @@ public struct ChatChannelView<Factory: ViewFactory>: View, KeyboardReadable {
                     .opacity(0) // Fixes showing accessibility button shape
                 }
                 .overlay(
-                    viewModel.reactionsShown ?
+                    viewModel.currentSnapshot != nil && messageDisplayInfo != nil && viewModel.reactionsShown ?
                         factory.makeReactionsOverlayView(
                             options: ReactionsOverlayViewOptions(
                                 channel: channel,