+++ Sources/StreamChatSwiftUI/ViewFactory/DefaultViewFactory.swift
@@ -834,6 +834,15 @@ extension ViewFactory {
         )
     }
     
+    public func makeQuotedMessageContentView(
+            options: QuotedMessageContentViewOptions
+        ) -> some View {
+            QuotedMessageContentView(
+                factory: self,
+                options: options
+            )
+        }
+    
     public func makeCustomAttachmentQuotedView(options: CustomAttachmentQuotedViewOptions) -> some View {
         EmptyView()
     }