+++ Sources/StreamChatSwiftUI/ViewFactory/ViewFactory.swift
@@ -513,6 +513,18 @@ import SwiftUI
     /// - Parameter options: the options for creating the quoted message view.
     func makeQuotedMessageView(options: QuotedMessageViewOptions) -> QuotedMessageViewType
     
+    associatedtype QuotedMessageContentViewType: View
+    /// Creates the quoted message content view.
+    ///
+    /// It is the view that is embedded in quoted message bubble view.
+    ///
+    /// - Parameters:
+    ///  - options: configuration options for the quoted message content view.
+    /// - Returns: view displayed in the quoted message content slot.
+    func makeQuotedMessageContentView(
+        options: QuotedMessageContentViewOptions
+    ) -> QuotedMessageContentViewType
+    
     associatedtype CustomAttachmentQuotedViewType: View
     /// Creates a quoted view for custom attachments. Returns `EmptyView` by default.
     /// - Parameter options: the options for creating the custom attachment quoted view.